+++ common/src/main/java/net/fabric_extras/ranged_weapon/api/BowMechanics.java
@@ -1,4 +1,0 @@
-package net.fabric_extras.ranged_weapon.api;
-
-public class BowMechanics {
-}