+++ common/src/main/java/net/fabric_extras/ranged_weapon/api/CustomCrossbow.java
@@ -1,6 +1,6 @@
 package net.fabric_extras.ranged_weapon.api;
 
-import net.fabric_extras.ranged_weapon.internal.RangedItemSettings;
+import net.fabric_extras.ranged_weapon.internal.AttributeUtils;
 import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.recipe.Ingredient;
@@ -12,12 +12,20 @@ public class CustomCrossbow extends CrossbowItem {
     // Instances are kept a list of, so model predicates can be automatically registered
     public final static HashSet<CustomCrossbow> instances = new HashSet<>();
 
-    public CustomCrossbow(Settings settings, RangedConfig config, Supplier<Ingredient> repairIngredientSupplier) {
+    public CustomCrossbow(Settings settings, RangedWeaponConfig config, Supplier<Ingredient> repairIngredientSupplier) {
         super(
-                ((RangedItemSettings)settings).rangedAttributes(config)
+                AttributeUtils.configure(settings, config)
         );
         this.repairIngredientSupplier = repairIngredientSupplier;
         instances.add(this);
+    }
+
+    /**
+     * @deprecated Use the {@link RangedWeaponConfig} constructor instead
+     */
+    @Deprecated
+    public CustomCrossbow(Settings settings, RangedConfig config, Supplier<Ingredient> repairIngredientSupplier) {
+        this(settings, config.toAbsolute(), repairIngredientSupplier);
     }
 
     private final Supplier<Ingredient> repairIngredientSupplier;